refactor: user info component moved
